Trait Peng::Planner

pub(crate) trait Planner {
    // Required methods
    fn plan(
        &self,
        current_position: Vector3<f32>,
        current_velocity: Vector3<f32>,
        time: f32,
    ) -> (Vector3<f32>, Vector3<f32>, f32);
    fn is_finished(&self, current_position: Vector3<f32>, time: f32) -> bool;
}

Trait defining the interface for trajectory planners

Required Methods

fn plan( &self, current_position: Vector3<f32>, current_velocity: Vector3<f32>, time: f32, ) -> (Vector3<f32>, Vector3<f32>, f32)

Plans the trajectory based on the current state and time

Arguments

current_position - The current position of the quadrotor

current_velocity - The current velocity of the quadrotor

time - The current simulation time

Returns

A tuple containing the desired position, velocity, and yaw angle

fn is_finished(&self, current_position: Vector3<f32>, time: f32) -> bool

Checks if the current trajectory is finished

Arguments

current_position - The current position of the quadrotor

time - The current simulation time

Returns

true if the trajectory is finished, false otherwise